forms renamed html_forms for diambiguation
Mathics3 Forms and web template forms are very different.
